+++ src/Scenario/Transformer/IsClickableTransformer.php
@@ -28,10 +28,11 @@ class IsClickableTransformer extends AbstractTransformer {
     protected function transform() : void
     {   
         $key = array_keys($this->step['isClickable'])[0];
+        $errorMessage = sprintf('%s[%s] is not clickable', $key, $this->step['isClickable'][$key]);
         $this->driver->wait(Configuration::get('wait.for'),Configuration::get('wait.every'))
                      ->until(WebDriverExpectedCondition::elementToBeClickable(
                         WebLocator::get($key, $this->step['isClickable'][$key])
-                    ));
+                    ), $errorMessage);
     }
 
     /**